full_examples/image_processing/requirements.txt
chalkpy[runtime]
beautifulsoup4
requests
pillow
CairoSVG==2.5.2
